Add File Counts markdown file
As a task for Day 9 of CS-340 Sp 2023, students will practice checking
out a repository and submitting changes with a pull request.  This file
gives students a place to submit a modification to a file.  On each linuxlab
machine, students will count the number of files in their home directory
and update the table accordingly.