Updates the Exam 1 Study Guide
Updates the Exam 1 Study Guide to match the level of progress currently
made. This includes an outline of all the topics to be covered in the
exam that Kevin shared with us during the lecturs. Additionally, it in-
cludes the content from the first two topics. This will be further
filled out and there is an outline of the future content to be added.