Update the Exam 1 Study Guide
This commit updates the exam 1 study guide to reflect the requests.
Those changes including fixixng misspellings that are in the document,
reformating definitions, adding code as inline text, and creating
sections for large code chunks. However, this commit does not put all
definitions as block quotes because this is the study guide document and
and including all the defintions in block qoutes becomes messy very
quickly.