Edits the Lecture 12 Notes
This commit updates the lecture 12 notes document. The first change that
was made to this document is changing the name of the file to align with
the style guide. The second change to this file is the addition of an
overview section at the begining of the document. The remaining changes
were making minor formating changes to better align with the style
guide.