Delete Day9Notes.md
This commit removes Day9Notes.md since it is a duplicate file

As noted in issue #163 the main repo currently contains two files : Day9Notes.md and Day12Notes.md that contain the same contents. After reviewing the files and checking with @hnsengimana Day9Notes.md is incorrectly labeled and should be removed to avoid confusion.